% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversor.R
\name{centimetros_a_pulgadas}
\alias{centimetros_a_pulgadas}
\title{convierte centimetros a pulgadas}
\usage{
centimetros_a_pulgadas(centimetros)
}
\arguments{
\item{centimetros}{Un valor o vector numerico en centimetros}
}
\value{
un valor o vector numerico en pulgadas
}
\description{
toma como entrada un valor o vector numerico en centimetros y devuelve su equivalente en pulgadas
}
\examples{
centimetros_a_pulgadas(10) # devuelve 3.937008

}
