% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma dos números}
\usage{
suma(x = 10, y = 20)
}
\arguments{
\item{x}{Un valor o vector numerico}

\item{y}{Un valor o vector numerico}
}
\value{
La suma de x e y
}
\description{
Suma dos números que dében ser positivos
}
\examples{
suma(5, 15)
suma(2, 3)

}
